scripts/requirements.txt
GitPython

